sie_module02/requirements.txt
wikipedia-api
python-dateutil 